Refactoring permutation generation and combining with test environment creation
