--- src/PerformanceTests/Common/Runner/AppDomains/AssemblyRedirectInfo.cs
@@ -1,11 +1,0 @@
-﻿namespace Utils.Runner.AppDomains
-{
-    using System;
-
-    public class AssemblyRedirectInfo
-    {
-        public string ShortName;
-        public Version Version;
-        public string PublicKeyToken;
-    }
-}
