--- src/PerformanceTests/Common/Runner/Nuget/Package.cs
@@ -1,13 +1,0 @@
-﻿namespace Utils.Runner.Nuget
-{
-    using System.IO;
-
-    public class Package
-    {
-        public PackageInfo Info { get; set; }
-
-        public string Version { get; set; }
-
-        public FileInfo[] Files { get; set; }
-    }
-}
