--- src/PerformanceTests/Common/Runner/Nuget/PackageVersionComparer.cs
@@ -1,19 +1,0 @@
-﻿namespace Utils.Runner.Nuget
-{
-    using System.Collections.Generic;
-
-    public class PackageVersionComparer : IEqualityComparer<Package>
-    {
-        public static PackageVersionComparer Default => new PackageVersionComparer();
-
-        public bool Equals(Package x, Package y)
-        {
-            return x.Version == y.Version;
-        }
-
-        public int GetHashCode(Package obj)
-        {
-            return obj.Version.GetHashCode();
-        }
-    }
-}
