+++ src/PerformanceTests/Common/Utils.csproj
@@ -70,15 +70,6 @@
     <Compile Include="BusCreationOptions.cs" />
     <Compile Include="NCrunchEnvironment.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Runner\AppDomains\TestEnvironmentCreator.cs" />
-    <Compile Include="Runner\AppDomains\AppDomainRunner.cs" />
-    <Compile Include="Runner\AppDomains\AssemblyLoader.cs" />
-    <Compile Include="Runner\AppDomains\AssemblyRedirectInfo.cs" />
-    <Compile Include="Runner\AppDomains\AssemblyRedirector.cs" />
-    <Compile Include="Runner\Nuget\PackageVersionComparer.cs" />
-    <Compile Include="Runner\OutputDirectoryCreator.cs" />
-    <Compile Include="Runner\Nuget\Package.cs" />
-    <Compile Include="Runner\Nuget\PackageInfo.cs" />
     <Compile Include="Statistics.cs" />
   </ItemGroup>
   <ItemGroup>
@@ -95,6 +86,9 @@
       <Isolated>False</Isolated>
       <EmbedInteropTypes>True</EmbedInteropTypes>
     </COMReference>
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Runner\" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
