+++ src/PerformanceTests/NServiceBus5/Program.cs
@@ -2,10 +2,10 @@ namespace NServiceBus5
 {
     using System;
     using System.Linq;
-    using Categories;
     using Common;
     using Common.Saga;
     using NServiceBus;
+    using Tests.Permutations;
     using Utils;
 
     class Program
