+++ src/PerformanceTests/NServiceBus6/ConcurrencyProfile.cs
@@ -1,10 +1,9 @@
-using Categories;
 using NServiceBus;
+using Tests.Permutations;
 
 class ConcurrencyProfile : IProfile, INeedPermutation
 {
     public Permutation Permutation { private get; set; }
-
 
     public void Configure(EndpointConfiguration cfg)
     {
