+++ src/PerformanceTests/PerformanceTests.sln
@@ -15,35 +15,35 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NServiceBus5", "NServiceBus
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NServiceBus6", "NServiceBus6\NServiceBus6.csproj", "{10D11044-C92C-42AD-830E-57012FF4F039}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Persistence.V5.NHibernate", "Persistence.V5.NHibernate.V6\Persistence.V5.NHibernate.csproj", "{207E1813-EC15-4C3B-ADBC-88E3733257E4}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Persistence.V5.NHibernate_v6", "Persistence.V5.NHibernate.V6\Persistence.V5.NHibernate_v6.csproj", "{207E1813-EC15-4C3B-ADBC-88E3733257E4}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Persistence.V6.NHibernate", "Persistence.V6.NHibernate.V7\Persistence.V6.NHibernate.csproj", "{02CA6A55-A771-4AD5-912A-484A2BBEFBD5}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Persistence.V6.NHibernate_v7", "Persistence.V6.NHibernate.V7\Persistence.V6.NHibernate_v7.csproj", "{02CA6A55-A771-4AD5-912A-484A2BBEFBD5}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Persistence.V5.RavenDB", "Persistence.V5.RavenDB.V3\Persistence.V5.RavenDB.csproj", "{2951D082-F37D-45F1-92FA-19DCFF3D8622}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Persistence.V5.RavenDB_v3", "Persistence.V5.RavenDB.V3\Persistence.V5.RavenDB_v3.csproj", "{2951D082-F37D-45F1-92FA-19DCFF3D8622}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Persistence.V6.RavenDB", "Persistence.V6.RavenDB.V4\Persistence.V6.RavenDB.csproj", "{C3E92DC8-F3C3-4CBF-BBC8-02B6DD440817}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Persistence.V6.RavenDB_v4", "Persistence.V6.RavenDB.V4\Persistence.V6.RavenDB_v4.csproj", "{C3E92DC8-F3C3-4CBF-BBC8-02B6DD440817}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Persistence.V5.InMemory", "Persistence.V5.InMemory.V5\Persistence.V5.InMemory.csproj", "{D8CB2DA9-4AE4-417C-89C3-00555F5E1F47}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Transport.V5.MSMQ.V5", "Transport.V5.MSMQ.V5\Transport.V5.MSMQ.V5.csproj", "{1C7DB98E-AEC5-4921-A765-AC81F1B8E40A}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Transport.V5.MSMQ", "Transport.V5.MSMQ.V5\Transport.V5.MSMQ.csproj", "{1C7DB98E-AEC5-4921-A765-AC81F1B8E40A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Transport.V6.MSMQ.V6", "Transport.V6.MSMQ.V6\Transport.V6.MSMQ.V6.csproj", "{182F83FE-F6C2-4B6C-BF65-48EDB28EA14A}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Transport.V6.MSMQ", "Transport.V6.MSMQ.V6\Transport.V6.MSMQ.csproj", "{182F83FE-F6C2-4B6C-BF65-48EDB28EA14A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Transport.V5.SQLServer.V2", "Transport.V5.SQLServer.V2\Transport.V5.SQLServer.V2.csproj", "{1017C202-2EA7-4074-9BF9-276921403F3D}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Transport.V5.SQLServer_v2", "Transport.V5.SQLServer.V2\Transport.V5.SQLServer_v2.csproj", "{1017C202-2EA7-4074-9BF9-276921403F3D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Transport.V6.SQLServer.V3", "Transport.V6.SQLServer.V3\Transport.V6.SQLServer.V3.csproj", "{DCECEB8B-70F6-459F-8893-57218338D0F7}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Transport.V6.SQLServer_v3", "Transport.V6.SQLServer.V3\Transport.V6.SQLServer_v3.csproj", "{DCECEB8B-70F6-459F-8893-57218338D0F7}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Transport.V5.RabbitMQ.V3", "Transport.V5.RabbitMQ.V3\Transport.V5.RabbitMQ.V3.csproj", "{87B17C90-2562-4187-A152-3000DD9A6E1B}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Transport.V5.RabbitMQ_v3", "Transport.V5.RabbitMQ.V3\Transport.V5.RabbitMQ_v3.csproj", "{87B17C90-2562-4187-A152-3000DD9A6E1B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Transport.V6.RabbitMQ.V4", "Transport.V6.RabbitMQ.V4\Transport.V6.RabbitMQ.V4.csproj", "{245211CB-E830-454E-BE75-DCFA904EBB8B}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Transport.V6.RabbitMQ_v4", "Transport.V6.RabbitMQ.V4\Transport.V6.RabbitMQ_v4.csproj", "{245211CB-E830-454E-BE75-DCFA904EBB8B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Transport.V5.AzureServiceBus", "Transport.V5.AzureServiceBus.V6\Transport.V5.AzureServiceBus.csproj", "{19BE6CC3-9715-4CFA-8B19-5CF42CAAE4A2}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Transport.V5.AzureServiceBus_v6", "Transport.V5.AzureServiceBus.V6\Transport.V5.AzureServiceBus_v6.csproj", "{19BE6CC3-9715-4CFA-8B19-5CF42CAAE4A2}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Transport.V6.AzureServiceBus.V7", "Transport.V6.AzureServiceBus.V7\Transport.V6.AzureServiceBus.V7.csproj", "{3F77A6DF-9683-4450-A0FE-45259461947F}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Transport.V6.AzureServiceBus_v7", "Transport.V6.AzureServiceBus.V7\Transport.V6.AzureServiceBus_v7.csproj", "{3F77A6DF-9683-4450-A0FE-45259461947F}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Transport.V5.AzureStorageQueues", "Transport.V5.AzureStorageQueues.V6\Transport.V5.AzureStorageQueues.csproj", "{152520F2-7D43-443B-B929-73BDFE9CE792}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Transport.V5.AzureStorageQueues_v6", "Transport.V5.AzureStorageQueues.V6\Transport.V5.AzureStorageQueues_v6.csproj", "{152520F2-7D43-443B-B929-73BDFE9CE792}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Transport.V6.AzureStorageQueues.V7", "Transport.V6.AzureStorageQueues.V7\Transport.V6.AzureStorageQueues.V7.csproj", "{B8CD6B76-6416-4C10-8746-EC65CA09B191}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Transport.V6.AzureStorageQueues_7", "Transport.V6.AzureStorageQueues.V7\Transport.V6.AzureStorageQueues_7.csproj", "{B8CD6B76-6416-4C10-8746-EC65CA09B191}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Persistence", "Persistence", "{3CB57DB2-A0E9-4C56-B3D4-9051F34BD2C1}"
 EndProject
