+++ src/PerformanceTests/Tests/Categories/OutboxVsDtcFixture.cs
@@ -2,36 +2,41 @@ namespace Categories
 {
     using System.Collections.Generic;
     using NUnit.Framework;
+    using Tests.Permutations;
+    using Tests.Tools;
     using Variables;
 
     [TestFixture(Description = "Outbox vs DTC")]
     public class OutboxVsDtcFixture : Base
     {
-        [Test]
-        [TestCaseSource(nameof(Generate))]
+        [TestCaseSource(typeof(OutboxVsDtc), nameof(TestEnvironment.Generate))]
         public override void SendLocal(Permutation permutation)
         {
             base.SendLocal(permutation);
         }
 
-        IEnumerable<Permutation> Generate()
+        class OutboxVsDtc : TestEnvironment
         {
-            return PermutationGenerator.Generate(new Permutations
+            protected override IEnumerable<Permutation> CreatePermutations()
             {
-                Versions = new[] { NServiceBusVersion.V5, NServiceBusVersion.V6, },
-                IOPS = new[] { IOPS.Default },
-                Platforms = new[] { Platform.x86, },
-                GarbageCollectors = new[] { GarbageCollector.Client, },
-                Transports = new[] { Transport.MSMQ },
-                Persisters = new[] { Persistence.NHibernate, Persistence.RavenDB, },
-                Serializers = new[] { Serialization.Json, },
-                MessageSizes = new[] { MessageSize.Tiny, MessageSize.Small, MessageSize.Medium, MessageSize.Large, },
-                OutboxModes = new[] { Outbox.Off, Outbox.On, },
-                DTCModes = new[] { DTC.Off, DTC.On, },
-                TransactionMode = new[] { TransactionMode.Default, },
-                AuditModes = new[] { Audit.Off },
-                ConcurrencyLevels = new[] { ConcurrencyLevel.EnvCores }
-            });
+                return PermutationGenerator.Generate(new Permutations
+                {
+                    Versions = new[] { NServiceBusVersion.V5, NServiceBusVersion.V6, },
+                    IOPS = new[] { IOPS.Default },
+                    Platforms = new[] { Platform.x86, },
+                    GarbageCollectors = new[] { GarbageCollector.Client, },
+                    Transports = new[] { Transport.MSMQ },
+                    Persisters = new[] { Persistence.NHibernate, Persistence.RavenDB, },
+                    Serializers = new[] { Serialization.Json, },
+                    MessageSizes = new[] { MessageSize.Tiny, MessageSize.Small, MessageSize.Medium, MessageSize.Large, },
+                    OutboxModes = new[] { Outbox.Off, Outbox.On, },
+                    DTCModes = new[] { DTC.Off, DTC.On, },
+                    TransactionMode = new[] { TransactionMode.Default, },
+                    AuditModes = new[] { Audit.Off },
+                    ConcurrencyLevels = new[] { ConcurrencyLevel.EnvCores }
+                });
+
+            }
         }
     }
 }
