+++ src/PerformanceTests/Tests/Categories/ResourceUtilizationFixture.cs
@@ -2,36 +2,40 @@ namespace Categories
 {
     using System.Collections.Generic;
     using NUnit.Framework;
+    using Tests.Permutations;
+    using Tests.Tools;
     using Variables;
 
     [TestFixture]
     public class ResourceUtilizationFixture : Base
     {
-        [Test]
-        [TestCaseSource(nameof(Generate))]
+        [TestCaseSource(typeof(ResourceUtilization), nameof(TestEnvironment.Generate))]
         public override void SendLocal(Permutation permutation)
         {
             base.SendLocal(permutation);
         }
 
-        IEnumerable<Permutation> Generate()
+        class ResourceUtilization : TestEnvironment
         {
-            return PermutationGenerator.Generate(new Permutations
+            protected override IEnumerable<Permutation> CreatePermutations()
             {
-                Versions = new[] { NServiceBusVersion.V5, NServiceBusVersion.V6, },
-                IOPS = new[] { IOPS.Default },
-                Platforms = new[] { Platform.x86, Platform.x64, },
-                GarbageCollectors = new[] { GarbageCollector.Client,GarbageCollector.Server,  },
-                Transports = new[] { Transport.MSMQ },
-                Persisters = new[] { Persistence.InMemory },
-                Serializers = new[] { Serialization.Json, },
-                MessageSizes = new[] { MessageSize.Tiny, },
-                OutboxModes = new[] { Outbox.Off },
-                DTCModes = new[] { DTC.On, },
-                TransactionMode = new[] { TransactionMode.Default, },
-                AuditModes = new[] { Audit.On, },
-                ConcurrencyLevels = new[] { ConcurrencyLevel.EnvCores }
-            });
+                return PermutationGenerator.Generate(new Permutations
+                {
+                    Versions = new[] { NServiceBusVersion.V5, NServiceBusVersion.V6, },
+                    IOPS = new[] { IOPS.Default },
+                    Platforms = new[] { Platform.x86, Platform.x64, },
+                    GarbageCollectors = new[] { GarbageCollector.Client, GarbageCollector.Server, },
+                    Transports = new[] { Transport.MSMQ },
+                    Persisters = new[] { Persistence.InMemory },
+                    Serializers = new[] { Serialization.Json, },
+                    MessageSizes = new[] { MessageSize.Tiny, },
+                    OutboxModes = new[] { Outbox.Off },
+                    DTCModes = new[] { DTC.On, },
+                    TransactionMode = new[] { TransactionMode.Default, },
+                    AuditModes = new[] { Audit.On, },
+                    ConcurrencyLevels = new[] { ConcurrencyLevel.EnvCores }
+                });
+            }
         }
     }
 }
