+++ src/PerformanceTests/Tests/Permutations/Permutations.cs
@@ -1,4 +1,4 @@
-﻿namespace Categories
+namespace Tests.Permutations
 {
     public class Permutations
     {
