+++ src/PerformanceTests/Tests/Tests.csproj
@@ -62,19 +62,19 @@
     <Compile Include="Categories\AuditOnVsOffFixture.cs" />
     <Compile Include="Categories\Base.cs" />
     <Compile Include="Categories\OutboxVsDtcFixture.cs" />
-    <Compile Include="Categories\Permutation.cs" />
-    <Compile Include="Categories\PermutationGenerator.cs" />
-    <Compile Include="Categories\PermutationParser.cs" />
-    <Compile Include="Categories\Permutations.cs" />
     <Compile Include="Categories\PersistersFixture.cs" />
     <Compile Include="Categories\PublishVsSendFixture.cs" />
     <Compile Include="Categories\ResourceUtilizationFixture.cs" />
     <Compile Include="Integration\MutatorsTests.cs" />
     <Compile Include="Integration\PermutationDirectoryResolverTests.cs" />
     <Compile Include="Integration\SagasTests.cs" />
+    <Compile Include="Permutations\Permutation.cs" />
+    <Compile Include="Permutations\PermutationGenerator.cs" />
+    <Compile Include="Permutations\PermutationParser.cs" />
+    <Compile Include="Permutations\Permutations.cs" />
     <Compile Include="Tools\PermutationDirectoryResolver.cs" />
-    <Compile Include="Tools\TestEnvironmentCreator.cs" />
-    <Compile Include="Tools\TestInfo.cs" />
+    <Compile Include="Tools\TestDescriptor.cs" />
+    <Compile Include="Tools\TestEnvironment.cs" />
     <Compile Include="Tools\Extensions.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Tools\TestsGlobal.cs" />
