+++ src/PerformanceTests/Tests/Tools/TestDescriptor.cs
@@ -1,0 +1,10 @@
+namespace Tests.Tools
+{
+    using Tests.Permutations;
+
+    public class TestDescriptor
+    {
+        public Permutation Permutation { get; set; }
+        public string ProjectAssemblyPath { get; set; }
+    }
+}
