+++ src/PerformanceTests/Transport.V6.MSMQ.V6/MsmqProfile.cs
@@ -1,6 +1,6 @@
 ﻿
-using Categories;
 using NServiceBus;
+using Tests.Permutations;
 using Variables;
 
 class MsmqProfile : IProfile, INeedPermutation
